Refactor and some docs
