+++ fastcrypto-vdf/benches/class_group.rs
@@ -29,7 +29,7 @@ fn class_group_ops_single<M: Measurement>(
 }
 
 fn class_group_ops(c: &mut Criterion) {
-    let mut group: BenchmarkGroup<_> = c.benchmark_group(format!("Class Group"));
+    let mut group: BenchmarkGroup<_> = c.benchmark_group("Class Group".to_string());
 
     class_group_ops_single("-9458193260787340859710210783898414376413627187338129653105774703043377776905956484932486183722303201135571583745806165441941755833466966188398807387661571", & mut group);
     class_group_ops_single("-173197108158285529655099692042166386683260486655764503111574151459397279244340625070436917386670107433539464870917173822190635872887684166173874718269704667936351650895772937202272326332043347073303124000059154982400685660701006453457007094026343973435157790533480400962985543272080923974737725172126369794019", &mut group);
